+++ modules/core/src/main/java/org/approvej/review/ReviewResultRecord.java
@@ -1,9 +1,9 @@
 package org.approvej.review;
 
 /**
- * Result class for {@link FileReviewer}s.
+ * Result class for {@link Reviewer}s.
  *
  * @param needsReapproval indicates if the result should be reapproved after the review (e.g.
  *     because the approval file was modified)
  */
-public record FileReviewResult(boolean needsReapproval) implements ReviewResult {}
+public record ReviewResultRecord(boolean needsReapproval) implements ReviewResult {}